refactor: Make all cli.admin module names singular
