+++ src/ai/backend/client/cli/admin/agent.py
