+++ src/ai/backend/client/cli/admin/group.py
