+++ src/ai/backend/client/cli/admin/image.py
