+++ src/ai/backend/client/cli/admin/keypair.py
