+++ src/ai/backend/client/cli/admin/resource.py
