+++ src/ai/backend/client/cli/admin/scaling_group.py
