+++ src/ai/backend/client/cli/admin/user.py
